Solved the n_tries problem. It is because the data type was set to double incorrectly.

+++ src/vars.h
@@ -270,7 +270,7 @@ double roff2_minus_ron2_cube; ///< used for switch potential
 
 /* set up parameters for this simulated annealing run */
 /* how many points do we try before stepping (move to the next temperature) */
-double n_tries;
+int n_tries;
 /* how many iterations for each T? (how many MD steps before use the acceptance critirea) */
 int iters_fixed_t;
 /* initial temperature */
